src/gurubase-backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
django-filter==23.3
django-extensions==3.2.3
gunicorn==22.0.0
Markdown==3.6
pylint==3.1.0
pylint-django==2.5.5
autopep8==2.3.1
nose==1.3.7
requests==2.32.3
responses==0.25.0
ipython==8.22.2
tqdm==4.66.4
pyyaml==6.0.1
uritemplate==4.1.1
coverage==7.4.4
python-decouple==3.8
psycopg2-binary==2.9.9
whitenoise==6.6.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
sentry-sdk==2.9.0
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-db-connection-retrier==1.0
faker==26.0.0
langchain==0.3.7
langchain-community==0.3.5
pymilvus==2.4.10
openai==1.57.0
bs4==0.0.2
typesense==0.21.0
anthropic==0.40.0
google-generativeai==0.8.3
redis==5.0.4
flower==2.0.1
pydantic==2.8.2
pydantic_core==2.20.1
pillow==10.4.0
django-storages==1.14.4
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
cffi==1.17.1
cryptography==43.0.1
pycparser==2.22
youtube_transcript_api==0.6.2
pytube2==15.0.11
pypdf==5.0.0
firecrawl-py==1.4.0
python-slugify==8.0.4
colorthief==0.2.1
CairoSVG==2.7.1
matplotlib==3.9.2
python-jose[cryptography]==3.3.0
GitPython==3.1.42
